tar: Rename entrypoint to tar.py
